// SDKs/twilio.com/twilio_proxy_v1/1.37.3/typescript/src/sdk/models/shared/proxyv1servicesessionparticipant.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class ProxyV1ServiceSessionParticipant extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_deleted" })
  dateDeleted?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @SpeakeasyMetadata({ data: "json, name=identifier" })
  identifier?: string;

  @SpeakeasyMetadata({ data: "json, name=links" })
  links?: Map<string, any>;

  @SpeakeasyMetadata({ data: "json, name=proxy_identifier" })
  proxyIdentifier?: string;

  @SpeakeasyMetadata({ data: "json, name=proxy_identifier_sid" })
  proxyIdentifierSid?: string;

  @SpeakeasyMetadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @SpeakeasyMetadata({ data: "json, name=session_sid" })
  sessionSid?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}